==> Extended_Figure6/03_DEGs_and_EnrichmentAnalysis/01A_DEG_Filter_Genes_D7_Y2020-2021.ipynb <==
import pandas as pd 
import numpy as np 
import scanpy as sc
import matplotlib.pyplot as plt
import concurrent.futures
import pickle
import warnings
from datetime import date
import hisepy
import os
import shutil
from concurrent.futures import ThreadPoolExecutor, as_completed,ProcessPoolExecutor
from tqdm import tqdm
import anndata
import gc
warnings.filterwarnings("ignore")
sc.settings.n_jobs = 60
print("Current working directory:", os.getcwd())

# Read MetaData

meta_data=pd.read_csv("/home/jupyter/BRI_Figures_Final_V2/Figure5/01_Frequency_Comparison/selected_samples_with_acutal_flu_year.csv")

# Assemble Data

meta_data_subset=meta_data[meta_data['Flu_Year'].isin(['2020-2021','2021-2022'])]
meta_data_subset=meta_data_subset[meta_data_subset['sample.visitName'].isin(['Flu Year 1 Day 7','Flu Year 2 Day 7'])]

meta_data_subset=meta_data_subset[meta_data_subset['Covid_exclusion']=='no']

%%time
file_names= ['/home/jupyter/BRI_Figures_Final_V2/Dataset/scRNA/BRI/h5ad/sample_h5ad/'+x+".h5ad" for x in meta_data_subset['pbmc_sample_id'].tolist()]
adata_list = []
with ThreadPoolExecutor(max_workers=30) as executor:
    future_to_file = {executor.submit(sc.read_h5ad, file_name): file_name for file_name in file_names}
    for future in tqdm(as_completed(future_to_file), total=len(file_names)):
        result = future.result()
        if result is not None:
            adata_list.append(result)

adata = anndata.concat(adata_list)
del adata_list 
gc.collect()

# Read Files and filter genes for DESEQ2

fitlered_gene=pd.DataFrame()
for i in adata.obs['AIFI_L3'].unique():
    print(i)
    adata_subset=adata[adata.obs['AIFI_L3']==i]
    sc.pp.filter_genes(adata_subset, min_cells=round(np.shape(adata_subset.X)[0]*0.1))
    gene_list=pd.DataFrame(list(adata_subset.var.index))
    gene_list.columns=['gene']
    gene_list['AIFI_L3']=i
    fitlered_gene=pd.concat([fitlered_gene,gene_list])

fitlered_gene.to_csv("filtered_gene_Y2020-2021_Y2021-2022_D7.csv")
