==> Extended_Figure6/03_DEGs_and_EnrichmentAnalysis/01B_DEG_analysis_D7_Y2020-2021.ipynb <==
suppressPackageStartupMessages(library(dplyr))
suppressPackageStartupMessages(library(tidyr))
suppressPackageStartupMessages(library(tibble))
suppressPackageStartupMessages(library(stringr))
library(parallel)
library(DESeq2)
library(hise)
source('/home//jupyter/BRI_Figures_Final_V2/helper_function/helper_function_IHA.r')

# Read Meta Data

meta_data=read.csv("/home/jupyter/BRI_Figures_Final_V2/Figure5/01_Frequency_Comparison/selected_samples_with_acutal_flu_year.csv")

# Perform Deseq2 on Year 1 Day 0 samples

filtered_gene_set<-read.csv("filtered_gene_Y2020-2021_Y2021-2022_D7.csv")

meta_data_subset=meta_data %>% filter(Flu_Day=='Day 7')

aggregated_count_file_list<-paste0("/home/jupyter/BRI_Figures_Final_V2/Dataset/scRNA/BRI/Aggregated_Count/Aggregated_Raw_Expression_of_Celltypes_by_Sample_AIFI_L3/",meta_data_subset$pbmc_sample_id,".csv")

df_list<-read_pseudobulk_expression(aggregated_count_file_list)

celltypes <- unique(unlist(lapply(df_list, names)))
celltypes<-unique(sub(".*:", "", celltypes))

res_list<-mclapply(celltypes, function(celltype){
    print(celltype)
    celltype_list <- lapply(df_list, function(df) { df <- df[, grep(celltype, names(df), fixed = TRUE),drop=FALSE] })
    exp_matrix<-do.call(cbind,celltype_list)
    colnames(exp_matrix)<-sub(":.*", "", colnames(exp_matrix))
    rownames(meta_data_subset)<-meta_data_subset$pbmc_sample_id
    filtered_gene_set_filtered<-filtered_gene_set%>% filter(AIFI_L3==celltype)%>% select(gene) %>%pull()
    
    res=deseq2_analysis(exp_matrix,
                     meta_data=meta_data_subset,
                     filtered_gene_set=filtered_gene_set_filtered,
                     formula= ~  cohort.cohortGuid+subject.biologicalSex+CMV+Flu_Year,
                     comparisons=list(c("cohort.cohortGuid", "BR1", "BR2")),#cooksCutoff=FALSE, independentFiltering=FALSE,
                     celltype=celltype)
    return(res)
},mc.cores=60)

res_all<-do.call(rbind,res_list)

head(res_all)

write.csv(res_all,"Deseq2_Result_Y2020-2021_Y2021-2022_D7.csv")
