==> Extended_Figure6/03_DEGs_and_EnrichmentAnalysis/05_Extract_Genes_CoreB_by_Sample_to_be_delete.ipynb <==
suppressPackageStartupMessages(library(dplyr))
suppressPackageStartupMessages(library(tidyr))
suppressPackageStartupMessages(library(tibble))
suppressPackageStartupMessages(library(stringr))
library(parallel)
library(rstatix)
library(data.table)
source('/home//jupyter/BRI_Figures_Final_V2/helper_function/helper_function_IHA.r')

meta_data=read.csv("/home/jupyter/BRI_Figures_Final_V2/Figure5/01_Frequency_Comparison/selected_samples_with_acutal_flu_year.csv")

file_list<-paste0("/home/jupyter/BRI_Figures_Final_V2/Dataset/scRNA/BRI/Average_LogNormalized_Expression/Average_LogNormalized_Expression_of_Celltypes_by_Sample_AIFI_L3/",meta_data$pbmc_sample_id,".csv")

df_list<-read_pseudobulk_expression(file_list)

genes_of_interest <- c( "IGHG1", "IGHG2","IGHG3", "IGHG4","IGHM","IGHD")

  
celltype_list <- lapply(df_list, function(df) { df <- df[genes_of_interest, grep('CD27- effector B cell', names(df), fixed = TRUE),drop=FALSE] })


exp_matrix<-do.call(cbind,celltype_list)

colnames(exp_matrix)<-sub(":.*", "", colnames(exp_matrix))
rownames(meta_data)<-meta_data$pbmc_sample_id

meta_data<-meta_data %>% filter(Flu_Year=='2020-2021',Flu_Day=='Day 7')

exp_matrix<-exp_matrix[meta_data$pbmc_sample_id]

    exp_matrix$RowNames <- rownames(exp_matrix)
    long_format <- melt(exp_matrix, id.vars = "RowNames")
    colnames(long_format) <- c('Gene', 'pbmc_sample_id', 'Mean_Expression')


long_format<-left_join(long_format,meta_data) 

colnames(long_format)

long_format %>% group_by(Gene)%>% wilcox_test(Mean_Expression~cohort.cohortGuid)

write.csv(long_format,"IGHG1_IGHG2_CoreB.csv")
